+++ tests/test_recognize.py
@@ -1047,7 +1047,7 @@ def test_wildcard() -> None:
     assert result is None, f"{sentence} should not match"
 
     # Test without text at the end
-    sentence = "start the white album by the beatles"
+    sentence = "start the white album by the beatles."
     result = recognize(sentence, intents)
     assert result is not None, f"{sentence} should match"
     assert set(result.entities.keys()) == {"album", "artist"}
@@ -1749,3 +1749,188 @@ def test_range_list_with_one_number() -> None:
     # Check ranges
     assert recognize("test 1", intents)
     assert not recognize("test 2", intents)
+
+
+def test_range_list_with_halves() -> None:
+    """Test a range list with fractions (1/2)."""
+    yaml_text = """
+    language: "en"
+    intents:
+      TestIntent:
+        data:
+          - sentences:
+              - "test {value}"
+    lists:
+      value:
+        range:
+          from: 1
+          to: 2
+          fractions: halves
+    """
+
+    with io.StringIO(yaml_text) as test_file:
+        intents = Intents.from_yaml(test_file)
+
+    assert recognize("test 1", intents)
+    result = recognize("test 1.5", intents)
+    assert result is not None
+    value = result.entities.get("value")
+    assert value is not None
+    assert value.value == 1.5
+    assert value.text == "1.5"
+    assert value.text_clean == "1.5"
+
+    result = recognize("test one point five", intents, language="en")
+    assert result is not None
+    value = result.entities.get("value")
+    assert value is not None
+    assert value.value == 1.5
+    assert value.text == "one point five"
+    assert value.text_clean == "one point five"
+
+    result = recognize("test 2.0", intents)
+    assert result is not None
+    value = result.entities.get("value")
+    assert value is not None
+    assert value.value == 2
+    assert value.text == "2.0"
+    assert value.text_clean == "2.0"
+
+    result = recognize("test 2.5", intents)
+    assert result is not None
+    value = result.entities.get("value")
+    assert value is not None
+    assert value.value == 2.5
+    assert value.text == "2.5"
+    assert value.text_clean == "2.5"
+
+    # Only halves
+    assert not recognize("test 2.1", intents)
+
+
+def test_range_list_with_tenths() -> None:
+    """Test a range list with fractions (1/10)."""
+    yaml_text = """
+    language: "en"
+    intents:
+      TestIntent:
+        data:
+          - sentences:
+              - "test {value}"
+    lists:
+      value:
+        range:
+          from: 1
+          to: 2
+          fractions: tenths
+    """
+
+    with io.StringIO(yaml_text) as test_file:
+        intents = Intents.from_yaml(test_file)
+
+    assert recognize("test 1", intents)
+    assert recognize("test 2", intents)
+
+    for integer in range(1, 3):
+        for tenth in range(1, 10):
+            value_str = f"{integer}.{tenth}"
+            result = recognize(f"test {value_str}", intents)
+            assert result is not None
+            value = result.entities.get("value")
+            assert value is not None
+            assert value.value == float(value_str)
+            assert value.text == value_str
+            assert value.text_clean == value_str
+
+    result = recognize("test one point one", intents, language="en")
+    assert result is not None
+    value = result.entities.get("value")
+    assert value is not None
+    assert value.value == 1.1
+    assert value.text == "one point one"
+
+    result = recognize("test two point six", intents, language="en")
+    assert result is not None
+    value = result.entities.get("value")
+    assert value is not None
+    assert value.value == 2.6
+    assert value.text == "two point six"
+
+    # Only tenths
+    assert not recognize("test 2.12", intents)
+
+
+def test_range_lists_separated_by_punctuation() -> None:
+    """Test range lists separated by punctuation."""
+    yaml_text = """
+    language: "en"
+    intents:
+      TestIntent:
+        data:
+          - sentences:
+              - "test {value1}.{value2}"
+    lists:
+      value1:
+        range:
+          from: 0
+          to: 1
+      value2:
+        range:
+          from: 0
+          to: 2
+    """
+
+    with io.StringIO(yaml_text) as test_file:
+        intents = Intents.from_yaml(test_file)
+
+    result = recognize("test 1.2", intents)
+    assert result is not None
+
+    value1 = result.entities.get("value1")
+    assert value1 is not None
+    assert value1.value == 1
+
+    value2 = result.entities.get("value2")
+    assert value2 is not None
+    assert value2.value == 2
+
+
+def test_range_lists_separated_by_punctuation_with_wildcard() -> None:
+    """Test range lists separated by punctuation preceeded by a wildcard."""
+    yaml_text = """
+    language: "en"
+    intents:
+      TestIntent:
+        data:
+          - sentences:
+              - "test {anything} {value1}.{value2}"
+    lists:
+      anything:
+        wildcard: true
+      value1:
+        range:
+          from: 0
+          to: 1
+      value2:
+        range:
+          from: 0
+          to: 2
+    """
+
+    with io.StringIO(yaml_text) as test_file:
+        intents = Intents.from_yaml(test_file)
+
+    result = recognize("test for the big 1.2", intents)
+    assert result is not None
+
+    anything = result.entities.get("anything")
+    assert anything is not None
+    assert anything.value == "for the big"
+
+    value1 = result.entities.get("value1")
+    assert value1 is not None
+    assert value1.value == 1
+
+    value2 = result.entities.get("value2")
+    assert value2 is not None
+    assert value2.value == 2